+++ packages/vendure-plugin-store-credit/src/api/api-extensions.ts
@@ -75,6 +75,7 @@ export const adminApiExtensions = gql`
   input CreateWalletInput {
     customerId: ID
     name: String!
+    code: String
     metadata: JSON
   }